bash: Enable gnu89 style
Apparently bash build uses very old style C functions with non-explicit
numbers of arguments.

THe source was some mkbuild.c file (or something).

Target only needed BUILD_FOR_CC but the host needed it for CC.

There's probably more to the story here, but this is fairly minimal.
(Maybe CC is not needed for HOST.)