grub: Disable string-null-termination copy warning
GCC15 raises a warning if a string is copied without its null
terminator, e.g. the literal "abcd" includes a NUL character which does
not fit in the four-character array `a`:

  char a[4] = "abcd"

GRUB includes a literal token of this form, and raises a warning, which
ROCKNIX converts to an error.

This patch disables the error in GRUB.  As with all `-W-no-*` disabled
warnings, this has no effect on older GCC compilers, and is safe to use.